Group connect list types in a single module
